2004-09-02  Bradford Hovinen  <hovinen@vega.pc.cis.udel.edu>
	* linbox/blackbox/diagonal.h (class Diagonal): Compilation fix

	* linbox/blackbox/dense.h (class DenseMatrix): Make _AT use
	DenseMatrix, not DenseMatrixBase

	* tests/test-la-block-lanczos.C (main):

	* tests/test-mg-block-lanczos.C (main): Use better defaults

	* linbox/algorithms/la-block-lanczos.inl (solve, sampleNullspace):
	Clear history when done

	* tests/Makefile.am (BASIC_TESTS): Add test-mg-block-lanczos

	* linbox/algorithms/la-block-lanczos.h (class
	LABlockLanczosSolver::InnerProductArray): Add reset

	* linbox/algorithms/la-block-lanczos.inl (iterate): Reset uAv

2004-07-06  Bradford Hovinen  <bghovinen@math.uwaterloo.ca>

	* linbox/blackbox/dense.h (class DenseMatrix): Add _AT, a
	TransposeMatrix of this

	* linbox/blackbox/dense.inl (DenseMatrix::applyTranspose): Use _AT

	* linbox/matrix/matrix-domain.h (class MatrixDomain): Get rid of
	template argument MatrixTrait on specialized permute

	* linbox/vector/vector-domain.h (class VectorDomain): Get rid of
	template argument VectorTrait on specialized permute, swap

	* linbox/matrix/sparse.h (struct MatrixTraits):

	* linbox/vector/vector-domain.inl (class VectorDomain):

	* linbox/vector/vector-domain.h (class VectorDomain):

	* linbox/matrix/matrix-domain.h (class MatrixDomain):

	* linbox/matrix/transpose.h (struct MatrixTraits):

	* linbox/matrix/dense.h (struct MatrixTraits): Remove templates
	from matrix tags

+++ linbox/linbox/field/modular.h
@@ -251,7 +251,7 @@ namespace LinBox
 		 * @param  x   field base element.
 		 */
 		std::ostream &write (std::ostream &os, const Element &x) const
-			{ return os << x; }
+			{ return os << (int) x; }
  
 
 		/*- Read field base element.
